+++ splunk_connect_for_snmp/common/collection_manager.py
@@ -1,10 +1,20 @@
+import logging
 import os
+import sys
 from abc import abstractmethod
 from contextlib import suppress
 
 import yaml
 from celery.utils.log import get_task_logger
+from jsonschema import ValidationError, validate
 
+from splunk_connect_for_snmp.common.collections_schemas import (
+    get_all_group_schemas,
+    get_all_profile_schemas,
+)
+from splunk_connect_for_snmp.common.customised_json_formatter import (
+    CustomisedJSONFormatter,
+)
 from splunk_connect_for_snmp.common.hummanbool import human_bool
 
 with suppress(ImportError, OSError):
@@ -14,7 +24,18 @@ with suppress(ImportError, OSError):
 
 CONFIG_PATH = os.getenv("CONFIG_PATH", "/app/config/config.yaml")
 CONFIG_FROM_MONGO = human_bool(os.getenv("CONFIG_FROM_MONGO", "false").lower())
-logger = get_task_logger(__name__)
+celery_logger = get_task_logger(__name__)
+
+log_level = "INFO"
+formatter = CustomisedJSONFormatter()
+logger = logging.getLogger(__name__)
+logger.setLevel(log_level)
+
+# writing to stdout
+handler = logging.StreamHandler(sys.stdout)
+handler.setLevel(log_level)
+handler.setFormatter(formatter)
+logger.addHandler(handler)
 
 
 class CollectionManager:
@@ -25,6 +46,11 @@ class CollectionManager:
     @staticmethod
     @abstractmethod
     def gather_elements():
+        pass
+
+    @staticmethod
+    @abstractmethod
+    def validate_elements(elements: dict):
         pass
 
     def return_collection_once(self):
@@ -62,12 +88,37 @@ class CollectionManager:
         # check in case only header is present
         if all_elements is None:
             all_elements = {}
+        self.validate_elements(all_elements)
         self.update_collection(all_elements)
 
 
 class GroupsManager(CollectionManager):
     def __init__(self, mongo):
         super().__init__(mongo, "groups")
+
+    @staticmethod
+    @abstractmethod
+    def validate_elements(elements: dict):
+        schemas = get_all_group_schemas()
+        invalid_groups = []
+        for group_name, group_body in elements.items():
+            valid = False
+            for schema in schemas:
+                try:
+                    validate(group_body, schema)
+                    valid = True
+                    break
+                except ValidationError:
+                    continue
+            if not valid:
+                invalid_groups.append(group_name)
+        for group in invalid_groups:
+            del elements[group]
+        if invalid_groups:
+            logger.error(
+                f"The following groups have invalid configuration and won't be used: {invalid_groups}. Please check "
+                f"indentation and keywords spelling inside inside mentioned groups configuration."
+            )
 
     def gather_elements(self):
         groups = {}
@@ -82,13 +133,37 @@ class GroupsManager(CollectionManager):
                     if "groups" in config_runtime:
                         groups = config_runtime.get("groups", {})
             except FileNotFoundError:
-                logger.info(f"File: {CONFIG_PATH} not found")
+                celery_logger.info(f"File: {CONFIG_PATH} not found")
         return groups
 
 
 class ProfilesManager(CollectionManager):
     def __init__(self, mongo):
         super().__init__(mongo, "profiles")
+
+    @staticmethod
+    @abstractmethod
+    def validate_elements(elements: dict):
+        schemas = get_all_profile_schemas()
+        invalid_profiles = []
+        for profile_name, profile_body in elements.items():
+            valid = False
+            for schema in schemas:
+                try:
+                    validate(profile_body, schema)
+                    valid = True
+                    break
+                except ValidationError:
+                    continue
+            if not valid:
+                invalid_profiles.append(profile_name)
+        for profile in invalid_profiles:
+            del elements[profile]
+        if invalid_profiles:
+            logger.error(
+                f"The following profiles have invalid configuration and won't be used: {invalid_profiles}. Please check "
+                f"indentation and keywords spelling inside inside mentioned profiles configuration."
+            )
 
     def gather_elements(self):
         active_profiles = {}
@@ -100,7 +175,7 @@ class ProfilesManager(CollectionManager):
             if file.endswith("yaml"):
                 with open(os.path.join(pkg_path, file), encoding="utf-8") as of:
                     profiles = yaml.safe_load(of)
-                    logger.info(
+                    celery_logger.info(
                         f"loading {len(profiles.keys())} profiles from shared profile group {file}"
                     )
                     for key, profile in profiles.items():
@@ -112,7 +187,7 @@ class ProfilesManager(CollectionManager):
                 profile = pr[key]
                 if key in active_profiles:
                     if not profile.get("enabled", True):
-                        logger.info(f"disabling profile {key}")
+                        celery_logger.info(f"disabling profile {key}")
                         del active_profiles[key]
                     else:
                         active_profiles[key] = profile
@@ -124,18 +199,18 @@ class ProfilesManager(CollectionManager):
                     config_runtime = yaml.safe_load(file)
                     if "profiles" in config_runtime:
                         profiles = config_runtime.get("profiles", {})
-                        logger.info(
+                        celery_logger.info(
                             f"loading {len(profiles.keys())} profiles from runtime profile group"
                         )
                         for key, profile in profiles.items():
                             if key in active_profiles:
                                 if not profile.get("enabled", True):
-                                    logger.info(f"disabling profile {key}")
+                                    celery_logger.info(f"disabling profile {key}")
                                     del active_profiles[key]
                                 else:
                                     active_profiles[key] = profile
                             else:
                                 active_profiles[key] = profile
             except FileNotFoundError:
-                logger.info(f"File: {CONFIG_PATH} not found")
+                celery_logger.info(f"File: {CONFIG_PATH} not found")
         return active_profiles